Moved OperationQueue and OperationProcessor into their own header + source files
