+++ src/vsg/io/read.cpp
@@ -14,6 +14,8 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 #include <vsg/io/ReaderWriter_vsg.h>
 #include <vsg/io/read.h>
 
+#include <vsg/threading/OperationProcessor.h>
+
 #include <thread>
 #include <iostream>
 #include <chrono>
@@ -62,211 +64,6 @@ ref_ptr<Object> vsg::read(const Path& filename, ref_ptr<const Options> options)
     return object;
 }
 
-namespace vsg
-{
-
-struct Active : public Object
-{
-    Active() : active(true) {}
-
-    std::atomic_bool active;
-
-    explicit operator bool() const noexcept { return active; }
-
-protected:
-    virtual ~Active() {}
-};
-
-
-struct Latch : public Object
-{
-    Latch(uint32_t num) :
-        count(num) {}
-
-    void count_down()
-    {
-        --count;
-        if (is_ready())
-        {
-            std::unique_lock lock(_mutex);
-            cv.notify_all();
-        }
-    }
-
-    bool is_ready() const
-    {
-        return (count==0);
-    }
-
-    void wait()
-    {
-        // use while loop to return immediate when latch already released
-        // and to handle cases where the condition variable releases spuriously.
-        while (count != 0)
-        {
-            std::unique_lock lock(_mutex);
-            cv.wait(lock);
-        }
-    }
-
-    std::atomic_uint count;
-    std::mutex _mutex;
-    std::condition_variable cv;
-
-protected:
-    virtual ~Latch() {}
-};
-
-struct Operation : public Object
-{
-    virtual void run() = 0;
-};
-
-
-class OperationQueue : public Inherit<Object, OperationQueue>
-{
-public:
-    OperationQueue(ref_ptr<Active> in_active) :
-        _active(in_active) {}
-
-    std::mutex _mutex;
-    std::condition_variable _cv;
-    std::list<ref_ptr<Operation>> _queue;
-    ref_ptr<Active> _active;
-
-    void add(ref_ptr<Operation> operation)
-    {
-        std::unique_lock lock(_mutex);
-        _queue.emplace_back(operation);
-        _cv.notify_one();
-    }
-
-    template<typename Iterator>
-    void add(Iterator begin, Iterator end)
-    {
-        size_t numAdditions = 0;
-        std::unique_lock lock(_mutex);
-        for(auto itr = begin; itr != end; ++itr)
-        {
-            _queue.emplace_back(*itr);
-            ++numAdditions;
-        }
-
-        if (numAdditions==1) _cv.notify_one();
-        else if (numAdditions>1) _cv.notify_all();
-    }
-
-    ref_ptr<Operation> take()
-    {
-        std::unique_lock lock(_mutex);
-
-        if (_queue.empty()) return {};
-
-        ref_ptr<Operation> operation = _queue.front();
-
-        _queue.erase(_queue.begin());
-
-        return operation;
-    }
-
-    ref_ptr<Operation> take_when_avilable()
-    {
-        std::chrono::duration waitDuration = std::chrono::milliseconds(100);
-
-        std::unique_lock lock(_mutex);
-
-        // wait to the conditional variable signals that an operation has been added
-        while (_queue.empty() && *_active)
-        {
-            //std::cout<<"Waiting on condition variable"<<std::endl;
-            _cv.wait_for(lock, waitDuration);
-        }
-
-        // if the threads we are associated with should no longer running go for a quick exit and return nothing.
-        if (!*_active)
-        {
-            return {};
-        }
-
-        // remove and return the head of the queue
-        ref_ptr<Operation> operation = _queue.front();
-        _queue.erase(_queue.begin());
-        return operation;
-    }
-};
-VSG_type_name(vsg::OperationQueue)
-
-class OperationProcessor : public Inherit<Object, OperationQueue>
-{
-public:
-
-    OperationProcessor(uint32_t numThreads, ref_ptr<Active> in_active = {}) :
-        active(in_active)
-    {
-        if (!active) active  = new Active;
-        queue = new OperationQueue(active);
-
-        auto run = [](ref_ptr<OperationQueue> q, ref_ptr<Active> a)
-        {
-            while(*(a))
-            {
-                ref_ptr<Operation> operation = q->take_when_avilable();
-                if (operation) operation->run();
-            }
-        };
-
-        for(size_t i=0; i<numThreads; ++i)
-        {
-            threads.emplace_back(std::thread(run, std::ref(queue), std::ref(active)));
-        }
-    }
-
-    void add(ref_ptr<Operation> operation)
-    {
-        queue->add(operation);
-    }
-
-    template<typename Iterator>
-    void add(Iterator begin, Iterator end)
-    {
-        queue->add(begin, end);
-    }
-
-    /// use this thread to run operations till the queue is empty as well
-    /// this thread will consume and run operations in parallel with any threads associated with this OperationProcessor.
-    void run()
-    {
-        while(ref_ptr<Operation> operation = queue->take())
-        {
-            operation->run();
-        }
-    }
-
-    /// stop theads
-    void stop()
-    {
-        active->active = false;
-        for(auto& thread : threads)
-        {
-            thread.join();
-        }
-        threads.clear();
-    }
-
-    using Threads = std::list<std::thread>;
-    Threads threads;
-    ref_ptr<OperationQueue> queue;
-    ref_ptr<Active> active;
-
-protected:
-    virtual ~OperationProcessor()
-    {
-        stop();
-    }
-};
-VSG_type_name(vsg::OperationProcessor)
-
-}
 
 struct ReadOperation : public Operation
 {
